feat: add view models for quest review summaries
