+++ html/Kickback/Backend/Views/vQuestReviewDetail.php
@@ -1,0 +1,16 @@
+<?php
+declare(strict_types=1);
+
+namespace Kickback\Backend\Views;
+
+class vQuestReviewDetail
+{
+    public int $accountId;
+    public string $username;
+    public ?string $avatar;
+    public ?int $hostRating;
+    public ?int $questRating;
+    public ?string $message;
+}
+
+?>
